backend/requirements.txt
psycopg[binary]>=3.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
httpx>=0.27.0
pydantic>=2.8.0
pytest>=8.2.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-genai>=1.30.0
aiofiles==23.2.1
python-multipart==0.0.6
opencv-python==4.9.0.80
google-cloud-tasks>=2.16.0
google-auth>=2.36.0
aiohttp>=3.8.0
nest-asyncio>=1.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
